+++ source/platforms/platform.ts
@@ -103,7 +103,7 @@ export function getPlatformForEnv(env: Env, source: CISource, requireAuth = true
   }
 
   // They need to set the token up for GitHub actions to work
-  if (env["GITHUB_EVENT_TYPE"] && !env["GITHUB_TOKEN"]) {
+  if (env["GITHUB_EVENT_NAME"] && !env["GITHUB_TOKEN"]) {
     console.error(`You need to add GITHUB_TOKEN to your Danger action in the workflow:
 
     action "${env["GITHUB_ACTION"]}" {